Implemented movement with keyboard for Hero class

+++ src/game_lib/elements/living_element.py
@@ -11,7 +11,7 @@ class LivingElement(MovingElement):
         position: Tuple[int, int],
         image_paths: list[str],
         dimensions: Tuple[int, int],
-        base_speed: float,
+        base_speed: int,
         velocity: Tuple[int, int],
     ) -> None:
         """Initialize LivingElement instance."""
